patient_generator/requirements.txt
Django==5.1.6
psycopg2==2.9.10
python-decouple==3.8
